+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Change GE Proton version from GE-Proton10-27 to GE-Proton10-33.
 - Change upgrade to run in 'tmux' session.
 - Change the virtual console from Getty to KMSCON.
+- Add 'dmemcg-booster' to improve VRAM usage for games in Gamescope and Plasma sessions.
 
 ## 4.5.0 - The Ventoy Resize Update